refactor eccentricity reweighting code to separate module
